Remove this bogus setting; always use 10000 as the default for sys.setcheckinterval().  Use floating-point fractions internally.

+++ rpython/translator/stm/src_stm/stmgcintf.h
@@ -46,7 +46,7 @@ static inline long pypy_stm_get_atomic(void) {
 }
 long pypy_stm_enter_callback_call(void);
 void pypy_stm_leave_callback_call(long);
-void pypy_stm_set_transaction_length(long);
+void pypy_stm_set_transaction_length(double);
 void pypy_stm_perform_transaction(object_t *, int(object_t *, int));
 
 static inline int pypy_stm_should_break_transaction(void)
